debian: Enable the SQL module by default.
Add the SQL module to the default load list, so that when upgrading
from an old unsplit config, which uses sql, we won't break horribly.

Also use the new -q (--silent-module-load) option when starting
syslog-ng, so that we won't get unnecessary warnings when a module is
not found.

Thanks to Imre Lazar <imre@balabit.hu> for the bug report.