debian: Prepare for multiple version support.
In debian/control.d/control.in, append @EXTRA_VERSION@ to each
relevant package.

The debian/tools/update-control.sh script replaces it with an empty
string if the requested upstream major version matches the default one
(which is set to 3.3 now), or with "-${UPSTREAM_VERSION_MAJOR}", if it
does not.

So debian/tools/update-control.sh 3.4.0 would end up with -3.4 as
@EXTRA_VERSION@.